Add ColumnPredicateBuilder and make HivePartitionFilterBuilder and JdbcPredicateBuilder extend from it
